Move user-related methods from AnimeRepository to UserRepository
