+++ app/src/main/java/com/alexis/myanimecompanion/ui/profile/ProfileViewModelFactory.kt
@@ -3,11 +3,15 @@ package com.alexis.myanimecompanion.ui.profile
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
 import com.alexis.myanimecompanion.data.AnimeRepository
+import com.alexis.myanimecompanion.data.UserRepository
 
-class ProfileViewModelFactory(private val animeRepository: AnimeRepository) : ViewModelProvider.Factory {
+class ProfileViewModelFactory(
+    private val animeRepository: AnimeRepository,
+    private val userRepository: UserRepository
+) : ViewModelProvider.Factory {
     override fun <T : ViewModel> create(modelClass: Class<T>): T {
         if (modelClass.isAssignableFrom(ProfileViewModel::class.java)) {
-            return ProfileViewModel(animeRepository) as T
+            return ProfileViewModel(animeRepository, userRepository) as T
         }
         throw IllegalArgumentException("Unknown ViewModel class")
     }
